requirements.txt
numpy
matplotlib==3.7.5
scikit-image
pyyaml
tqdm
scipy
networkx
numba
shapely
pymunk==6.5.1
torch
torchvision
gymnasium[mujoco]
stable-baselines3
tensorboard
pygame
opencv-python
pynput
git+https://github.com/rdesc/pydubins.git
trimesh
mujoco
packcircles
mapbox-earcut
